02-azd-deploy-hosted-agent/src/my-hosted-agent/requirements.txt
# Agent Framework
agent-framework-core

# Azure AI Hosting Adapter
azure-ai-agentserver-agentframework

# Azure OpenAI
openai
azure-identity
